Reorganize the FreeBSD/DragonFly code.
There is no shared x86 and x86_64 code between the two anymore. Instead these 4
lines of code were duplicated. We now only share architecture-independet
code between FreeBSD and DragonFly.
