+++ src/unix/bsd/freebsdlike/dragonfly/mod.rs
@@ -1,0 +1,8 @@
+cfg_if! {
+    if #[cfg(target_arch = "x86_64")] {
+        mod x86_64;
+        pub use self::x86_64::*;
+    } else {
+        // ...
+    }
+}
