+++ src/unix/bsd/freebsdlike/dragonfly/x86_64.rs
@@ -1,3 +1,8 @@
+pub type c_long = i64;
+pub type c_ulong = u64;
+pub type time_t = i64;
+pub type suseconds_t = i64;
+
 pub type clock_t = u64;
 pub type ino_t = u64;
 pub type nlink_t = u32;
@@ -56,7 +61,6 @@ s!{
         pub ss_size: ::size_t,
         pub ss_flags: ::c_int,
     }
-
 }
 
 extern {
