+++ src/unix/bsd/freebsdlike/freebsd/x86.rs
@@ -1,3 +1,8 @@
+pub type c_long = i32;
+pub type c_ulong = u32;
+pub type time_t = i32;
+pub type suseconds_t = i32;
+
 s! {
     pub struct stat {
         pub st_dev: ::dev_t,
