+++ src/unix/bsd/freebsdlike/freebsd/x86_64.rs
@@ -1,3 +1,8 @@
+pub type c_long = i64;
+pub type c_ulong = u64;
+pub type time_t = i64;
+pub type suseconds_t = i64;
+
 s! {
     pub struct stat {
         pub st_dev: ::dev_t,
