--- src/unix/bsd/freebsdlike/x86.rs
@@ -1,4 +1,0 @@
-pub type c_long = i32;
-pub type c_ulong = u32;
-pub type time_t = i32;
-pub type suseconds_t = i32;
