--- src/unix/bsd/freebsdlike/x86_64.rs
@@ -1,4 +1,0 @@
-pub type c_long = i64;
-pub type c_ulong = u64;
-pub type time_t = i64;
-pub type suseconds_t = i64;
